fix sign-compare warning in _stdio_write
Use size_t for loop variable to match s->length type,
fixing -Werror=sign-compare failure on CI.

+++ spy/libspy/include/spy/__spy__.h
@@ -12,7 +12,7 @@ spy___spy__$is_compiled(void) {
 
 static inline void
 spy___spy__$_stdio_write(spy_Str *s) {
-    for (int32_t i = 0; i < s->length; i++)
+    for (size_t i = 0; i < s->length; i++)
         putchar(s->utf8[i]);
     putchar('\n');
 }
